requirements.txt
Flask==2.3.3
groq==0.4.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0